fix breadcrumb labels
